Refactor and cleanup
Rename card states
