+++ webapp/src/abstractions/index.ts
@@ -2,3 +2,4 @@
 
 export * from './Card';
 export * from './ControllerInterface';
+export * from './SeatOf';
