+++ webapp/src/abstractions/messages/KittyRequest.ts
@@ -1,9 +1,9 @@
-import { CardInfo } from "./CardInfo";
+import { Card } from "..";
 
 export class KittyRequest {
     constructor(
         public matchId: number,
         public playerId: number,
-        public cards: CardInfo[],
+        public cards: Card[],
     ) {}
 }
