+++ webapp/src/abstractions/messages/index.ts
@@ -1,6 +1,5 @@
 // This directory contains messages for sockets communication
 
-export * from './CardInfo'
 export * from './AlertResponse';
 export * from './BidRequest';
 export * from './BidResponse';
